Attempt to port GDB support from FreeRTOS-Amiga (on Dante's request).
