+++ .gdbinit
@@ -1,0 +1,8 @@
+# some generic settings
+set output-radix 0x10
+set pagination off
+set confirm off
+set verbose off
+
+source gdb-dashboard
+dashboard -layout breakpoints assembly registers source stack variables
